+++ sound-mind-studio/tests/test_main_window.h
@@ -181,4 +181,5 @@ private slots:
     void bringPickedObjectToFrontMovesItAboveLaterStrokesOnTheSameLayer();
     void editingAPickedStrokesPathMovesANodeAndCommitsOnApply();
     void cancelingAPickedStrokesPathEditDiscardsTheDragWithoutCommitting();
+    void cutRegionIsPickableAndMovable();
 };